validation_nautilus/requirements.txt
boto3>=1.28.0
pyarrow>=14.0.0
pandas>=2.2.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pyyaml>=6.0.1
click>=8.1.0
colorama>=0.4.6
jinja2>=3.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
tabulate>=0.9.0